Added option to set toolbar start position with either an array or a function returning an array. Added option to save user-dragged position of individual toolbars, and have the toolbar jump to it when the related element is targeted.
Added options to show / hide title tags, the title bar itself and the to set the title default content in cases where the title tags are disabled.

Added _dialog.alert.

Added exception and dialog alert when rangy is not present.

Added ability for the initialization function to cancel load - by returning false.